+++ media/lua/client/UI/CHC_uses_recipepanel.lua
@@ -16,6 +16,10 @@ local fhLarge = texMan:getFontHeight(UIFont.Large) -- largeFontHeight
 local fhMedium = texMan:getFontHeight(UIFont.Medium) -- mediumFontHeight
 local fhSmall = texMan:getFontHeight(UIFont.Small) -- smallFontHeight
 
+function CHC_uses_recipepanel:initialise()
+    ISPanel.initialise(self)
+    self.fastListReturn = CHC_main.common.fastListReturn
+end
 
 function CHC_uses_recipepanel:createChildren()
     ISPanel.createChildren(self);
@@ -85,6 +89,7 @@ function CHC_uses_recipepanel:createChildren()
     self.ingredientPanel.itemheight = fhSmall + 2 * self.itemMargin
     self.ingredientPanel.font = UIFont.NewSmall
     self.ingredientPanel.doDrawItem = self.drawIngredient
+    self.ingredientPanel.yScroll = 0
     self.ingredientPanel.drawBorder = true
     self.ingredientPanel.borderColor = listBorderColor
     self.ingredientPanel.vscroll.borderColor = listBorderColor
@@ -127,6 +132,7 @@ function CHC_uses_recipepanel:createChildren()
     -- self.skillPanel:setOnMouseDownFunction(self, self.onIngredientMouseDown)
     self.skillPanel.itemheight = fhSmall + 2 * self.itemMargin
     self.skillPanel.doDrawItem = self.drawSkill
+    self.skillPanel.yScroll = 0
     self.skillPanel.drawBorder = true
     self.skillPanel.borderColor = listBorderColor
     self.skillPanel.vscroll.borderColor = listBorderColor
@@ -141,6 +147,7 @@ function CHC_uses_recipepanel:createChildren()
     self.booksPanel:setOnMouseDownFunction(self, self.onIngredientMouseDown)
     self.booksPanel.itemheight = fhSmall + 2 * self.itemMargin
     self.booksPanel.doDrawItem = self.drawBook
+    self.booksPanel.yScroll = 0
     self.booksPanel.drawBorder = true
     self.booksPanel.borderColor = listBorderColor
     self.booksPanel.vscroll.borderColor = listBorderColor
@@ -155,6 +162,7 @@ function CHC_uses_recipepanel:createChildren()
     self.equipmentPanel:setOnMouseDownFunction(self, self.onIngredientMouseDown)
     self.equipmentPanel.itemheight = fhSmall + 2 * self.itemMargin
     self.equipmentPanel.doDrawItem = self.drawBook
+    self.equipmentPanel.yScroll = 0
     self.equipmentPanel.drawBorder = true
     self.equipmentPanel.borderColor = listBorderColor
     self.equipmentPanel.vscroll.borderColor = listBorderColor
@@ -257,7 +265,9 @@ function CHC_uses_recipepanel:getSources(recipe)
                 insert(sourceMain, item)
             end
         end
-        for order, source in ipairs({ sourceBase, sourceMain, sourceSpice }) do
+        local types = { sourceBase, sourceMain, sourceSpice }
+        for order = 1, #types do
+            local source = types[order]
             local sourceInList = {}
             sourceInList.items = {}
 
@@ -485,8 +495,6 @@ end
 -- region update
 
 function CHC_uses_recipepanel:refreshIngredientPanel()
-    self.ingredientPanel:setVisible(false)
-
     local selectedItem = self.newItem
     if not selectedItem then return end
 
@@ -497,21 +505,33 @@ function CHC_uses_recipepanel:refreshIngredientPanel()
     else
         selectedItem.typesAvailable = self:getAvailableItemsType()
     end
-
+    local c1 = not utils.areTablesDifferent(selectedItem.typesAvailable, self.lastAvailableTypes)
+    local c2 = self.lastAvailableTypes and not utils.empty(self.lastAvailableTypes)
+    if c1 and c2 then
+        return
+    end
+    self.parent.backRef.updateQueue:push({
+        targetView = self.parent.ui_type,
+        actions = { 'needUpdateTypes' }
+    })
+    self.lastAvailableTypes = selectedItem.typesAvailable
+    self.ingredientPanel:setVisible(false)
     self.ingredientPanel:clear()
 
     -- Display single-item sources before multi-item sources
     local sortedSources = {}
-    for _, source in ipairs(selectedItem.sources) do
-        insert(sortedSources, source)
+    for i = 1, #selectedItem.sources do
+        insert(sortedSources, selectedItem.sources[i])
     end
     table.sort(sortedSources, function(a, b) return #a.items == 1 and #b.items > 1 end)
 
-    for _, source in ipairs(sortedSources) do
+    for i = 1, #sortedSources do
+        local source = sortedSources[i]
         local available = {}
         local unavailable = {}
 
-        for _, item in ipairs(source.items) do
+        for j = 1, #source.items do
+            local item = source.items[j]
             local data = {}
             data.isDestroy = source.isDestroy
             data.isKeep = source.isKeep
@@ -571,20 +591,17 @@ function CHC_uses_recipepanel:refreshIngredientPanel()
         end
         --]]
 
-        for k, item in ipairs(available) do
-            self.ingredientPanel:addItem(item.name, item)
+        for j = 1, #available do
+            self.ingredientPanel:addItem(available[j].name, available[j])
         end
-        for k, item in ipairs(unavailable) do
-            self.ingredientPanel:addItem(item.name, item)
+        for j = 1, #unavailable do
+            self.ingredientPanel:addItem(unavailable[j].name, unavailable[j])
         end
     end
-
-    self.refreshTypesAvailableMS = getTimestampMs()
 
     local h = math.min(10, #self.ingredientPanel.items) * self.ingredientPanel.itemheight
     self.ingredientPanel.origH = h
     --self.ingredientPanel:setHeight(h)
-    self.ingredientPanel.doDrawItem = CHC_uses_recipepanel.drawIngredient
     self.ingredientPanel:setVisible(true)
 end
 
@@ -829,14 +846,14 @@ function CHC_uses_recipepanel:drawIngredient(y, item, alt)
             if item.item.multiple then
                 fr, fb = 0.5, 0.1
             end
-            self:drawRect(1, y, self:getWidth() - 2, self.itemheight, fa, fr, fg, fb)
+            self:drawRect(1, y, self.width - 2, self.itemheight, fa, fr, fg, fb)
         end
     end
     local ab, rb, gb, bb = 1, 0.1, 0.1, 0.1
     if item.item.multipleHeader then
-        self:drawRect(1, y, self:getWidth() - 2, self.itemheight, 0.2, 0.25, gb, bb)
+        self:drawRect(1, y, self.width - 2, self.itemheight, 0.2, 0.25, gb, bb)
     end
-    self:drawRectBorder(0, y, self:getWidth() - 2, self.itemheight, ab, rb, gb, bb)
+    self:drawRectBorder(0, y, self.width - 2, self.itemheight, ab, rb, gb, bb)
     -- ISUIElement:drawRectBorder( x, y, w, h, a, r, g, b)
 
     return y + self.itemheight;
@@ -855,7 +872,7 @@ function CHC_uses_recipepanel:drawSkill(y, item, alt)
         a = 0.7
     end
     self:drawText(text, 15, y, r, g, b, a, UIFont.Small)
-    self:drawRectBorder(0, y, self:getWidth() - 2, self.itemheight, ab, rb, gb, bb)
+    self:drawRectBorder(0, y, self.width - 2, self.itemheight, ab, rb, gb, bb)
     return y + self.itemheight
 end
 
@@ -882,7 +899,7 @@ function CHC_uses_recipepanel:drawBook(y, item, alt)
     self.parent.drawFavoriteStar(self, y, item)
     --endregion
 
-    self:drawRectBorder(0, y, self:getWidth() - 2, self.itemheight, ab, rb, gb, bb)
+    self:drawRectBorder(0, y, self.width - 2, self.itemheight, ab, rb, gb, bb)
 
     return y + self.itemheight
 end
@@ -942,25 +959,43 @@ end
 -- endregion
 
 function CHC_uses_recipepanel:render()
-    ISPanel.render(self);
+    ISPanel.render(self)
 
     if not self.recipe then return end
+    if self.needUpdateScroll then
+        self.ingredientPanel.yScroll = self.ingredientPanel:getYScroll()
+        self.skillPanel.yScroll = self.skillPanel:getYScroll()
+        self.booksPanel.yScroll = self.booksPanel:getYScroll()
+        self.equipmentPanel.yScroll = self.equipmentPanel:getYScroll()
+        self.needUpdateScroll = false
+    end
+
+    if self.needUpdateMousePos then
+        self.ingredientPanel.mouseX = self.ingredientPanel:getMouseX()
+        self.skillPanel.mouseX = self.skillPanel:getMouseX()
+        self.booksPanel.mouseX = self.booksPanel:getMouseX()
+        self.equipmentPanel.mouseX = self.equipmentPanel:getMouseX()
+        self.ingredientPanel.mouseY = self.ingredientPanel:getMouseY()
+        self.skillPanel.mouseY = self.skillPanel:getMouseY()
+        self.booksPanel.mouseY = self.booksPanel:getMouseY()
+        self.equipmentPanel.mouseY = self.equipmentPanel:getMouseY()
+        self.needUpdateMousePos = false
+    end
     local x = 10;
     local y = 10;
     local selectedItem = self.newItem
 
     -- region check if available
-    local now = getTimestampMs()
-
-    if not self.refreshTypesAvailableMS then
-        self.refreshTypesAvailableMS = now
-    end
-
-    if now > self.refreshTypesAvailableMS + 500 and self.needRefreshIngredientPanel == false then
+    local ms = UIManager.getMillisSinceLastRender()
+    if not self.ms then self.ms = 0 end
+    self.ms = self.ms + ms
+    if self.ms > 500 then
         self.needRefreshIngredientPanel = true
+        self.ms = 0
     end
 
     if self.needRefreshIngredientPanel then
+        self.needRefreshIngredientPanel = false
         local typesAvailable = self:getAvailableItemsType()
         self.needRefreshRecipeCounts = utils.areTablesDifferent(selectedItem.typesAvailable, typesAvailable)
         selectedItem.typesAvailable = typesAvailable
@@ -979,9 +1014,7 @@ function CHC_uses_recipepanel:render()
                 self.containerList, nil
             )
         end
-        self.refreshTypesAvailableMS = now
         self:refreshIngredientPanel()
-        self.needRefreshIngredientPanel = false
     end
 
     if self.needRefreshRecipeCounts then
@@ -1089,7 +1122,7 @@ function CHC_uses_recipepanel:drawCraftButtons(x, y, item)
     end
 
     if not self.craftAllButton:isVisible() and count > 1 then
-        self.craftAllButton:setX(self.craftOneButton:getX() + 5 + self.craftOneButton:getWidth())
+        self.craftAllButton:setX(self.craftOneButton.x + 5 + self.craftOneButton.width)
         self.craftAllButton:setVisible(true)
     end
     --endregion
@@ -1288,7 +1321,8 @@ function CHC_uses_recipepanel:craft(button, all)
     local container = itemsUsed[1]:getContainer()
     if not selectedItem.recipe:isCanBeDoneFromFloor() then
         container = self.player:getInventory()
-        for _, item in ipairs(itemsUsed) do
+        for i = 1, #itemsUsed do
+            local item = itemsUsed[i]
             if item:getContainer() ~= self.player:getInventory() then
                 insert(returnToContainer, item)
             end
@@ -1314,12 +1348,6 @@ end
 -- endregion
 
 
-function CHC_uses_recipepanel:fastListReturn(y)
-    if y + self:getYScroll() >= self.height then return true end
-    if y + self.itemheight + self:getYScroll() <= 0 then return true end
-    return false
-end
-
 -- endregion
 
 function CHC_uses_recipepanel:new(args)
@@ -1342,10 +1370,13 @@ function CHC_uses_recipepanel:new(args)
     o.playerNum = player and player:getPlayerNum() or -1
     o.needRefreshIngredientPanel = true
     o.needRefreshRecipeCounts = true
+    o.needUpdateScroll = false
+    o.needUpdateMousePos = false
     o.recipe = nil;
     o.manualsSize = 0
     o.manualsEntries = nil
     o.modData = CHC_main.playerModData
+    o.lastAvailableTypes = {}
 
     o.bh = nil